Enable --dry-run option

+++ main.go
@@ -4,11 +4,15 @@ import (
 	"crypto/sha1"
 	"encoding/hex"
 	"errors"
+	"flag"
+	"fmt"
 	"io"
 	"log"
 	"os"
 	"path/filepath"
 )
+
+var dryRun bool
 
 var (
 	errArgumentRequired              = errors.New("argument required")
@@ -16,21 +20,28 @@ var (
 )
 
 func main() {
-	if err := execute(os.Args[1:]); err != nil {
+	if err := execute(os.Args); err != nil {
 		log.Fatal(err)
 	}
 }
 
 func execute(args []string) error {
-	l := len(args)
+	flg := flag.NewFlagSet(args[0], flag.ExitOnError)
 
+	flg.BoolVar(&dryRun, "dry-run", false, "Print the execution result without executing it")
+
+	flg.Parse(args[1:])
+
+	ss := flg.Args()
+
+	l := len(ss)
 	if l == 0 {
 		return errArgumentRequired
 	} else if l > 1 {
 		return errOnlyOneArgumentCanBeSpecified
 	}
 
-	return filepath.Walk(args[0], walkFn)
+	return filepath.Walk(ss[0], walkFn)
 }
 
 func walkFn(path string, fi os.FileInfo, err error) error {
@@ -51,18 +62,22 @@ func rename(path string, fi os.FileInfo) error {
 		return err
 	}
 
-	_, err = os.OpenFile(newpath, os.O_CREATE|os.O_WRONLY|os.O_EXCL, 0644)
-	if err != nil {
-		if os.IsExist(err) {
-			return nil
+	if !dryRun {
+		_, err = os.OpenFile(newpath, os.O_CREATE|os.O_WRONLY|os.O_EXCL, 0644)
+		if err != nil {
+			if os.IsExist(err) {
+				return nil
+			}
+			return err
 		}
-		return err
+
+		err = os.Rename(path, newpath)
+		if err != nil {
+			return err
+		}
 	}
 
-	err = os.Rename(path, newpath)
-	if err != nil {
-		return err
-	}
+	fmt.Println(path + " -> " + newpath)
 
 	return nil
 }
